+++ pkg/providers/cloudaccess/testdata/TestDataKeyVersions/default.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudaccess_key_versions" "test" {
+  access_key_name = "Home automation | s3"
+}